requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
APScheduler==3.11.2
attrs==25.4.0
audioop-lts==0.2.2
discord.py==2.6.4
dnspython==2.8.0
frozenlist==1.8.0
idna==3.11
multidict==6.7.0
propcache==0.4.1
pymongo==4.16.0
pytz==2025.2
six==1.17.0
tzdata==2025.3
tzlocal==5.3.1
yarl==1.22.0
